Refactorisation de l'authentification et de la gestion des locations : mise à jour des types d'ID en Integer, simplification de la gestion des exceptions et amélioration de la lisibilité du code.

+++ Back-end/src/main/java/com/chatop/dtos/MessageDto.java
@@ -6,9 +6,9 @@ import lombok.Setter;
 @Getter
 @Setter
 public class MessageDto {
-    private Long id;
-    private Long rentalId;
-    private Long userId;
+    private Integer id;
+    private Integer rentalId;
+    private Integer userId;
     private String message;
     private String createdAt;
     private String updatedAt;
