+++ Back-end/src/main/java/com/chatop/repositories/RentalRepository.java
@@ -3,6 +3,6 @@ package com.chatop.repositories;
 import com.chatop.models.Rental;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RentalRepository extends JpaRepository<Rental, Long> {
+public interface RentalRepository extends JpaRepository<Rental, Integer> {
 
 }
